requirements.txt
aiohttp==3.13.2
pyyaml==6.0.3
sqlalchemy==2.0.44
asyncpg==0.30.0
aiohttp_swagger3==0.10.0
openapi-generator-cli[jdk4py]==7.17.0
pytest==8.4.2
pytest_aiohttp==1.1.0
flake8==7.3.0
flake8-quotes==3.4.0
